// src/generatedApis/models/reading.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Objects API
 * API for managing objects
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 *
 * @export
 * @interface Reading
 */
export interface Reading {
  /**
   *
   * @type {number}
   * @memberof Reading
   */
  readingId: number;
  /**
   *
   * @type {Date}
   * @memberof Reading
   */
  timestamp: Date;
  /**
   *
   * @type {number}
   * @memberof Reading
   */
  temperature: number;
  /**
   *
   * @type {number}
   * @memberof Reading
   */
  porePressure: number;
}